[c2goto] include libc headers in c2goto -> default to internal libc
