+++ src/c2goto/CMakeLists.txt
@@ -1,12 +1,24 @@
-add_executable(c2goto c2goto.cpp cprover_blank_library.cpp)
+
+add_subdirectory (headers)
+
+file(MAKE_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/c2goto-hdr)
+file(TOUCH ${CMAKE_CURRENT_BINARY_DIR}/c2goto-hdr/libc.h
+           ${CMAKE_CURRENT_BINARY_DIR}/c2goto-hdr/libm.h)
+file(MAKE_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/c2goto-hdr/headers)
+file(WRITE ${CMAKE_CURRENT_BINARY_DIR}/c2goto-hdr/headers/libc_hdr.h
+           "#include \"${CMAKE_CURRENT_BINARY_DIR}/headers/libc_hdr.h\"")
+
+add_executable(c2goto c2goto.cpp cprover_blank_library.cpp cprover_libc_sources.cpp)
 target_include_directories(c2goto
     PRIVATE ${Boost_INCLUDE_DIRS}
+    PRIVATE ${CMAKE_CURRENT_BINARY_DIR}/c2goto-hdr
     )
-  
 target_link_libraries(c2goto 
-  clangcfrontend gotoprograms langapi util_esbmc irep2
+  clangcfrontend gotoprograms langapi util_esbmc irep2 c2gotoheaders
   default_message ${Boost_LIBRARIES})
-add_subdirectory (headers)
+set_target_properties(c2goto PROPERTIES
+  PRIVATE_HEADER "${CMAKE_CURRENT_BINARY_DIR}/c2goto-hdr/libc.h;${CMAKE_CURRENT_BINARY_DIR}/c2goto-hdr/libm.h;${CMAKE_CURRENT_BINARY_DIR}/c2goto-hdr/headers/libc_hdr.h")
+
 
 if(DEFINED C2GOTO_INCLUDE_DIR)
   set(multiarch "${C2GOTO_INCLUDE_DIR}")
@@ -49,7 +61,7 @@ function(mangle_clib output)
     else()
       set(inputs_c ${c2goto_library_files})
     endif()
-    set(CMD c2goto -I ${CMAKE_CURRENT_SOURCE_DIR}/headers -I ${multiarch} ${OS_C2GOTO_FLAGS} ${inputs_c} ${in_flags} --output ${out_goto})
+    set(CMD c2goto -I ${multiarch} ${OS_C2GOTO_FLAGS} ${inputs_c} ${in_flags} --output ${out_goto})
     add_custom_command(OUTPUT ${out_goto}
       COMMAND ${CMD}
       DEPENDS c2goto ${c2goto_library_files} ${c2goto_libm_files} ${c2goto_header_files}
@@ -81,7 +93,7 @@ endfunction()
 
 message(STATUS "Multiarch include for c2goto: ${multiarch}. You may want to override C2GOTO_INCLUDE_DIR")
 mangle_clib(mangled_files ${inputs})
-add_library(clibs ${mangled_files} cprover_library.cpp)
+add_library(clibs ${mangled_files} cprover_library.cpp cprover_libc_sources.cpp)
 target_include_directories(clibs
     PRIVATE ${Boost_INCLUDE_DIRS}
     PRIVATE ${CMAKE_BINARY_DIR}/src
