+++ src/c2goto/cprover_blank_library.cpp
@@ -1,14 +1,7 @@
 
-#include <string>
-
 class contextt;
 class messaget;
 class languaget;
-
-const std::string *internal_libc_header_dir()
-{
-  return nullptr;
-}
 
 void add_cprover_library(contextt &, const messaget &, const languaget *)
 {
